+++ Sources/ShortcutFoundation/Core/Network/Core/Calls/DataToJSON.swift
@@ -1,0 +1,19 @@
+//
+//  NetworkingClient+JSON.swift
+//  ShortcutFoundation
+//
+//  Created by Gabriel Sabadin, Karl Söderberg on 2021-08-16.
+//  Copyright © 2021 Shortcut Scandinavia Apps AB. All rights reserved.
+//
+
+import Foundation
+import Combine
+
+// Data to JSON
+extension AnyPublisher where Output == Data {
+    public func toJSON() -> AnyPublisher<Any, Error> {
+         tryMap { data -> Any in
+            return try JSONSerialization.jsonObject(with: data, options: [])
+        }.eraseToAnyPublisher()
+    }
+}
