--- Sources/ShortcutFoundation/Core/Network/Core/Calls/NetworkingClient+Data.swift
@@ -1,54 +1,0 @@
-//
-//  NetworkingClient+Data.swift
-//  ShortcutFoundation
-//
-//  Created by Gabriel Sabadin, Karl Söderberg on 2021-08-16.
-//  Copyright © 2021 Shortcut Scandinavia Apps AB. All rights reserved.
-//
-
-import Foundation
-import Combine
-
-public extension NetworkingClient {
-    
-    
-//    func getData<Payload: Params>(_ route: String, params: Payload) -> AnyPublisher<Data, Error> {
-//        request(.get, route, params: params).publisher()
-//    }
-//    
-//    func getData(_ route: String) -> AnyPublisher<Data, Error> {
-//        request(.get, route).publisher()
-//    }
-//
-//    func postData<Payload: Params>(_ route: String, params: Payload) -> AnyPublisher<Data, Error> {
-//        request(.post, route, params: params).publisher()
-//    }
-//    
-//    func postData(_ route: String) -> AnyPublisher<Data, Error> {
-//        request(.post, route).publisher()
-//    }
-//
-//    func putData<Payload: Params>(_ route: String, params: Payload) -> AnyPublisher<Data, Error> {
-//        request(.put, route, params: params).publisher()
-//    }
-//    
-//    func putData(_ route: String) -> AnyPublisher<Data, Error> {
-//        request(.put, route).publisher()
-//    }
-//
-//    func patchData<Payload: Params>(_ route: String, params: Payload? = nil) -> AnyPublisher<Data, Error> {
-//        request(.patch, route, params: params).publisher()
-//    }
-//    
-//    func patchData(_ route: String) -> AnyPublisher<Data, Error> {
-//        request(.patch, route).publisher()
-//    }
-//
-//    func deleteData<Payload: Params>(_ route: String, params: Payload) -> AnyPublisher<Data, Error> {
-//        request(.delete, route, params: params).publisher()
-//    }
-//    
-//    func deleteData(_ route: String) -> AnyPublisher<Data, Error> {
-//        request(.delete, route).publisher()
-//    }
-}
